api: Move daemons to its own module
We are going to make new server/ui module and have the rest API as one
the modules under it. This is the first part of the move to achieve
this. By making this its own module, other code under this "ui" server
will be able to define their own daemons.

Signed-off-by: Andy Doan <andy@foundries.io>